projects/python/don-confiado-backend/app/requirements.txt
# FastAPI and web framework
fastapi[all]
fastapi-utils
httpx
pydantic
python-dotenv
uvicorn

# AI/ML dependencies
openai>=1.45.0
google-generativeai>=0.7.2

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=1.0.7
langchain-text-splitters>=0.1.0

# Graph database
neo4j>=5.24.0
neo4j-graphrag>=0.1.0

# Enhanced GraphRAG dependencies
# Note: neo4j-graphrag includes most required components
# Additional dependencies for advanced features
apoc>=0.1.0  # For APOC procedures in Neo4j

# Database
supabase>=2.6.0
SQLAlchemy
psycopg2-binary

# Document processing
pypdf>=4.3.1

# Utilities
typing-inspect
